moving attack helpers to their own file
